Update generated protobuf files with new protoc version
Regenerated files using protoc v6.33.2 from npm package.
